+++ src/test/java/com/cogpunk/math/probability/EventProbabilityProfileMultiplicationAggregationStrategyTest.java
@@ -2,7 +2,6 @@ package com.cogpunk.math.probability;
 
 import static org.junit.Assert.assertEquals;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 
 import org.junit.Test;
@@ -18,7 +17,7 @@ public class EventProbabilityProfileMultiplicationAggregationStrategyTest {
 
 		Integer[] args = new Integer[] {1,2,3,4,5,6};
 		
-		assertEquals((Integer) 720, new EventProbabilityProfileMultiplicationAggregationStrategy<Integer>(new IntegerOperator()).aggregate(Arrays.asList(args)));
+		assertEquals(720, new EventProbabilityProfileMultiplicationAggregationStrategy<Integer>(new IntegerOperator()).aggregate(Arrays.asList(args)));
 		
 		
 		//assertEquals((Integer) 0, new ProbabilityProfileMultiplicationAggregationStrategy<Integer>(new IntegerOperator()).aggregate(new ArrayList<Integer>()));
